ComponentProxyStrategy extracted to separate file.
