+++ GUI/coregui/Models/ProxyModelStrategy.h
@@ -63,21 +63,4 @@ protected:
                             const QModelIndex& index);
 };
 
-//! Strategy for ComponentProxyModel which hides extra level of GroupProperty.
-
-class BA_CORE_API_ ComponentProxyStrategy : public ProxyModelStrategy
-{
-public:
-    void onDataChanged(SessionModel* source, ComponentProxyModel* proxy);
-
-protected:
-    void processSourceIndex(SessionModel* model, ComponentProxyModel* proxy,
-                            const QModelIndex& index);
-private:
-    bool isGroupChildren(SessionItem* item);
-    void processGroupItem(SessionItem* item, const QPersistentModelIndex& sourceIndex,
-                          const QPersistentModelIndex& proxyIndex);
-
-};
-
 #endif  // ProxyModelStrategy
